Move logic to determine native types for OGR sources from OGR provider to a generic method in QgsOgrUtils
